# app/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Wardrobe Fitter API",
    "description": "API documentation for the Wardrobe Fitter application"
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/collection": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get all collections",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of collections"
          }
        }
      }
    },
...
    "/style": {
      "get": {
        "tags": ["Styles"],
        "summary": "Get all styles",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of styles"
          }
        }
      }
    },
    "/style/{style}": {
      "get": {
        "tags": ["Styles"],
        "summary": "Get specific style",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "style",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Style details"
          }
        }
      },
      "delete": {
        "tags": ["Styles"],
        "summary": "Delete a style",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "style",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Style deleted successfully"
          }
        }
      }
    },
    "/style/{style}/comments": {
      "get": {
        "tags": ["Styles"],
        "summary": "Get comments on a style",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "style",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments"
          }
        }
      }
    },
    "/item": {
      "get": {
        "tags": ["Items"],
        "summary": "Get all items",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of items"
          }
        }
      }
    },
...
    "/brand": {
      "get": {
        "tags": ["Brands"],
        "summary": "Get all brands",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of brands"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}